sfix solve s1_topic_6/2.2 (numeric types)

+++ training/section_1/topic_6/s1_t_6.cpp
@@ -170,7 +170,7 @@ vector<Document> SearchServer::FindAllDocuments(const Query &query) const {
 }
 
 int SearchServer::ComputeAverageRating(const vector<int> &ratings) {
-    const auto size = ratings.size();
+    const auto size = static_cast<int>(ratings.size());
     if (!size) return 0;
     const auto result =
         accumulate(ratings.begin(), ratings.end(), 0.0,
